:art: Rename model types file to activations
