+++ rnnalyse/activations/activation_writer.py
@@ -8,7 +8,7 @@ import numpy as np
 
 from ..typedefs.corpus import Labels
 from ..typedefs.extraction import ActivationRanges
-from ..typedefs.models import ActivationFiles, ActivationNames, PartialArrayDict
+from ..typedefs.activations import ActivationFiles, ActivationNames, PartialArrayDict
 from ..utils.paths import dump_pickle, trim
 from .activation_reader import ActivationReader
