+++ rnnalyse/interventions/mechanism.py
@@ -10,7 +10,7 @@ from typing import Callable, Tuple, Dict
 from torch import Tensor
 
 from ..models.intervention_lstm import InterventionLSTM
-from ..typedefs.models import FullActivationDict
+from ..typedefs.activations import FullActivationDict
 
 
 class InterventionMechanism(ABC):
