+++ rnnalyse/models/intervention_lstm.py
@@ -3,7 +3,7 @@ from typing import Dict
 from overrides import overrides
 
 from ..models.forward_lstm import ForwardLSTM
-from ..typedefs.models import FullActivationDict
+from ..typedefs.activations import FullActivationDict
 
 
 class InterventionLSTM(ForwardLSTM):
