+++ rnnalyse/typedefs/activations.py
@@ -1,4 +1,4 @@
-from typing import BinaryIO, Dict, List, Tuple, Union
+from typing import BinaryIO, Dict, List, Union, Tuple
 
 import numpy as np
 from torch import Tensor
@@ -8,6 +8,8 @@ ActivationNames = List[ActivationName]
 
 ActivationFiles = Dict[ActivationName, BinaryIO]
 
+ActivationIndex = Union[int, slice, List[int], np.ndarray]
+
 ActivationLayer = Dict[str, Tensor]
 
 # Nested dict with embeddings for each activation
@@ -16,6 +18,5 @@ FullActivationDict = Dict[int, ActivationLayer]
 # Dict with arbitrary number of activations
 PartialActivationDict = Dict[ActivationName, Tensor]
 PartialArrayDict = Dict[ActivationName, Union[np.ndarray, List[np.ndarray]]]
-
 
 ParameterDict = FullActivationDict
