+++ rnnalyse/typedefs/classifiers.py
@@ -2,7 +2,7 @@ from typing import Any, Dict
 
 import numpy as np
 
-from .models import ActivationName
+from .activations import ActivationName
 
 DataDict = Dict[str, np.ndarray]
 ResultsDict = Dict[ActivationName, Dict[str, Any]]
